Kulthuset position much more stable
I tried lots of scripts and things to attatch the Kulthuset prefab to a plane, as an anchor etc. to make it more stable and "float" around less. I added an anchor component to the prefab which maybe didnt make a difference. Once I set the prefab position and rotations to 0 (except 2 rotations) it's now much more stable for some reason...